Move mock_favorite_places to mocks folder.
