+++ src/tests/mocks/mock_favorite_places.py
